+++ generator/syntax_highlighting/syntax_highlighting.js
@@ -1,0 +1,2 @@
+exports.panther = require("./panther.js").highlight;
+exports.cpp = require("./cpp.js").highlight;